requirements.txt
fabric>=3.2.2
pyyaml>=6.0.1
pydantic>=2.7.0
datasets>=2.19.0
transformers>=4.41.0
accelerate>=0.30.1
peft>=0.11.1
huggingface-hub>=0.23.0
jsonschema>=4.21.1
rich>=13.7.1
unsloth>=2024.5.7
trl>=0.8.6
litellm>=1.43.12
